クラス名DisplayMemoをMemoDisplayerへ変更,HandleFileをMemoDisplayerへ統合
